+++ tests/test_cut.py
@@ -2,6 +2,7 @@ import unittest
 import topojson
 import geopandas
 from shapely import geometry
+from topojson.core.cut import Cut
 
 
 class TestCut(unittest.TestCase):
@@ -11,7 +12,7 @@ class TestCut(unittest.TestCase):
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "abc2": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
         }
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         # print(topo)
         self.assertEqual(len(topo["junctions"]), 0)
         self.assertSequenceEqual(topo["bookkeeping_duplicates"].tolist(), [[1, 0]])
@@ -22,7 +23,7 @@ class TestCut(unittest.TestCase):
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "cba": {"type": "LineString", "coordinates": [[2, 0], [1, 0], [0, 0]]},
         }
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         # print(topo)
         self.assertEqual(len(topo["junctions"]), 0)
         self.assertSequenceEqual(topo["bookkeeping_duplicates"].tolist(), [[1, 0]])
@@ -39,7 +40,7 @@ class TestCut(unittest.TestCase):
                 "coordinates": [[[0, 0], [1, 0], [2, 1], [0, 0]]],
             },
         }
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         # print(topo)
         self.assertEqual(len(topo["junctions"]), 0)
         self.assertSequenceEqual(topo["bookkeeping_duplicates"].tolist(), [[1, 0]])
@@ -56,7 +57,7 @@ class TestCut(unittest.TestCase):
                 "coordinates": [[[0, 0], [2, 1], [1, 0], [0, 0]]],
             },
         }
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         # print(topo)
         self.assertEqual(len(topo["junctions"]), 0)
         self.assertSequenceEqual(topo["bookkeeping_duplicates"].tolist(), [[1, 0]])
@@ -73,7 +74,7 @@ class TestCut(unittest.TestCase):
                 "coordinates": [[[1, 0], [2, 1], [0, 0], [1, 0]]],
             },
         }
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         # print(topo)
         self.assertEqual(len(topo["junctions"]), 0)
         self.assertSequenceEqual(topo["bookkeeping_duplicates"].tolist(), [[1, 0]])
@@ -90,7 +91,7 @@ class TestCut(unittest.TestCase):
                 "coordinates": [[[0, 0], [1, 0], [2, 1], [0, 0]]],
             },
         }
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         # print(topo)
         self.assertEqual(len(topo["junctions"]), 0)
         self.assertSequenceEqual(topo["bookkeeping_duplicates"].tolist(), [[1, 0]])
@@ -107,7 +108,7 @@ class TestCut(unittest.TestCase):
                 "coordinates": [[[1, 0], [2, 1], [0, 0], [1, 0]]],
             },
         }
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         # print(topo)
         self.assertEqual(len(topo["junctions"]), 0)
         self.assertSequenceEqual(topo["bookkeeping_duplicates"].tolist(), [[1, 0]])
@@ -124,7 +125,7 @@ class TestCut(unittest.TestCase):
                 "coordinates": [[[0, 0], [1, 0], [2, 1], [0, 0]]],
             },
         }
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         # print(topo)
         self.assertEqual(len(topo["junctions"]), 0)
         self.assertSequenceEqual(topo["bookkeeping_duplicates"].tolist(), [[1, 0]])
@@ -144,17 +145,17 @@ class TestCut(unittest.TestCase):
                 "coordinates": [[[1, 0], [2, 0], [2, 1], [1, 1], [1, 0]]],
             },
         }
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         # print(topo)
         self.assertEqual(topo["bookkeeping_linestrings"].size, 6)
         self.assertSequenceEqual(topo["bookkeeping_duplicates"].tolist(), [[4, 1]])
 
     # currently the border between Sudan and Eqypt is not recognized as duplicate
-    # because of floating-precision. Should be solved by a 'snap_to_grid' option.
+    # because of floating-precision. Should be solved by a "snap_to_grid" option.
     def test_to_cut_border_egypt_sudan(self):
         data = geopandas.read_file(geopandas.datasets.get_path("naturalearth_lowres"))
         data = data[(data.name == "Egypt") | (data.name == "Sudan")]
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         self.assertEqual(len(topo["bookkeeping_duplicates"].tolist()), 0)
 
     def test_nybb_fast_split(self):
@@ -162,7 +163,7 @@ class TestCut(unittest.TestCase):
         data = geopandas.read_file(nybb_path)
         data.set_index("BoroCode", inplace=True)
 
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         self.assertEqual(topo["bookkeeping_linestrings"].size, 5618)
 
     # this test was added since the fast_split was really slow on geometries
@@ -172,7 +173,7 @@ class TestCut(unittest.TestCase):
         data = geopandas.read_file(
             "tests/files_geojson/mesh2d.geojson", driver="GeoJSON"
         )
-        topo = topojson.Cut(data).to_dict()
+        topo = Cut(data).to_dict()
         self.assertEqual(topo["bookkeeping_linestrings"].size, 11010)
 
     def test_super_function_cut(self):
@@ -182,14 +183,15 @@ class TestCut(unittest.TestCase):
                 geometry.Polygon([[1, 0], [2, 0], [2, 1], [1, 1], [1, 0]]),
             ]
         )
-        topo = topojson.cut(data)
+        topo = Cut(data).to_dict()
         self.assertEqual(
-            list(topo.cut.keys()),
+            list(topo.keys()),
             [
                 "type",
                 "linestrings",
                 "bookkeeping_geoms",
                 "objects",
+                "options",
                 "junctions",
                 "bookkeeping_duplicates",
                 "bookkeeping_linestrings",
