+++ tests/test_hashmap.py
@@ -2,6 +2,7 @@ import unittest
 import topojson
 import geopandas
 from shapely import geometry
+from topojson.core.hashmap import Hashmap
 
 
 class TestHasmap(unittest.TestCase):
@@ -28,7 +29,7 @@ class TestHasmap(unittest.TestCase):
                 ],
             }
         }
-        topo = topojson.Hashmap(data).to_dict()
+        topo = Hashmap(data).to_dict()
         # print(topo)
         self.assertEqual(
             topo["objects"]["data"]["geometries"][0]["arcs"], [[[4, 0]], [[1]], [[2]]]
@@ -45,7 +46,7 @@ class TestHasmap(unittest.TestCase):
                 "coordinates": [[[1, 0], [2, 0], [2, 1], [1, 1], [1, 0]]],
             },
         }
-        topo = topojson.Hashmap(data).to_dict()
+        topo = Hashmap(data).to_dict()
         self.assertEqual(topo["objects"]["data"]["geometries"][0]["arcs"], [[-3, 0]])
         self.assertEqual(topo["objects"]["data"]["geometries"][1]["arcs"], [[1, 2]])
 
@@ -54,7 +55,7 @@ class TestHasmap(unittest.TestCase):
     def test_albania_greece(self):
         data = geopandas.read_file(geopandas.datasets.get_path("naturalearth_lowres"))
         data = data[(data.name == "Albania") | (data.name == "Greece")]
-        topo = topojson.Hashmap(data).to_dict()
+        topo = Hashmap(data).to_dict()
         self.assertEqual(len(topo["arcs"]), 3)
 
     # something is wrong with hashmapping in the example of benin
@@ -65,7 +66,7 @@ class TestHasmap(unittest.TestCase):
             | (data.name == "Benin")
             | (data.name == "Burkina Faso")
         ]
-        topo = topojson.Hashmap(data).to_dict()
+        topo = Hashmap(data).to_dict()
         self.assertEqual(len(topo["arcs"]), 6)
 
     # something is wrong with hashmapping once a geometry has only shared arcs
@@ -78,7 +79,7 @@ class TestHasmap(unittest.TestCase):
             | (data.name == "Namibia")
             | (data.name == "Zambia")
         ]
-        topo = topojson.Hashmap(data).to_dict()
+        topo = Hashmap(data).to_dict()
         self.assertEqual(len(topo["arcs"]), 14)
 
     # this test was added since the shared_arcs bookkeeping is doing well, but the
@@ -92,7 +93,7 @@ class TestHasmap(unittest.TestCase):
             | (data.name == "Mozambique")
             | (data.name == "Zambia")
         ]
-        topo = topojson.Hashmap(data).to_dict()
+        topo = Hashmap(data).to_dict()
         self.assertEqual(len(topo["arcs"]), 18)
 
     def test_super_function_hashmap(self):
@@ -102,5 +103,5 @@ class TestHasmap(unittest.TestCase):
                 geometry.Polygon([[1, 0], [2, 0], [2, 1], [1, 1], [1, 0]]),
             ]
         )
-        topo = topojson.Hashmap(data).to_dict()
-        self.assertEqual(list(topo.hashmap.keys()), ["type", "objects", "arcs"])
+        topo = Hashmap(data).to_dict()
+        self.assertEqual(list(topo.keys()), ["type", "objects", "options", "arcs"])
