+++ tests/test_join.py
@@ -2,6 +2,7 @@ from shapely import geometry
 import unittest
 import topojson
 import geopandas
+from topojson.core.join import Join
 
 
 class TestJoin(unittest.TestCase):
@@ -11,7 +12,7 @@ class TestJoin(unittest.TestCase):
             "cba": {"type": "LineString", "coordinates": [[2, 0], [1, 0], [0, 0]]},
             "ab": {"type": "LineString", "coordinates": [[0, 0], [1, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(geometry.MultiPoint([
             (0.0, 0.0),(1.0, 0.0),(2.0, 0.0)
             ]).equals(geometry.MultiPoint(topo["junctions"])))
@@ -22,7 +23,7 @@ class TestJoin(unittest.TestCase):
             "cba": {"type": "LineString", "coordinates": [[2, 0], [1, 0], [0, 0]]},
             "ab": {"type": "LineString", "coordinates": [[0, 0], [2, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         # print(topo)
         self.assertFalse(
             geometry.Point(1.0, 0.0) in geometry.MultiPoint(topo["junctions"])
@@ -41,7 +42,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [(5, 0), (30, 0), (30, 5), (0, 5)],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         # print(topo)
         self.assertTrue(len(topo["junctions"]), 4)
 
@@ -59,7 +60,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [(5, 0), (30, 0), (30, 5), (0, 5)],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         # print(topo)
         self.assertTrue(len(topo["junctions"]), 6)
 
@@ -69,7 +70,7 @@ class TestJoin(unittest.TestCase):
             "abc1": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "abc2": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         # print(topo)
         self.assertListEqual(topo["junctions"], [])
 
@@ -79,7 +80,7 @@ class TestJoin(unittest.TestCase):
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "cba": {"type": "LineString", "coordinates": [[2, 0], [1, 0], [0, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         # print(topo)
         self.assertListEqual(topo["junctions"], [])
 
@@ -95,7 +96,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[0, 0], [1, 1], [2, 0], [0, 0]]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         # print(topo)
         self.assertListEqual(topo["junctions"], [])
 
@@ -111,7 +112,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[0, 0], [2, 0], [1, 1], [0, 0]]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # rotated duplicate rings BCAB & ABCA have no junctions
@@ -126,7 +127,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[1, 1], [2, 0], [0, 0], [1, 1]]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # ring ABCA & line ABCA have no junction at A
@@ -141,7 +142,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[0, 0], [1, 1], [2, 0], [0, 0]]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # ring ABCA & line ABCA have no junctions
@@ -156,7 +157,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[1, 1], [2, 0], [0, 0], [1, 1]]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when an old arc ABC extends a new arc AB, there is a junction at B
@@ -165,7 +166,7 @@ class TestJoin(unittest.TestCase):
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "ab": {"type": "LineString", "coordinates": [[0, 0], [1, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(
             geometry.MultiPoint(topo["junctions"]).equals(
                 geometry.MultiPoint([(0.0, 0.0), (1.0, 0.0)])
@@ -178,7 +179,7 @@ class TestJoin(unittest.TestCase):
             "cba": {"type": "LineString", "coordinates": [[2, 0], [1, 0], [0, 0]]},
             "ab": {"type": "LineString", "coordinates": [[0, 0], [1, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(
             geometry.Point(1.0, 0.0).within(geometry.MultiPoint(topo["junctions"]))
         )
@@ -189,7 +190,7 @@ class TestJoin(unittest.TestCase):
             "ade": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 1], [2, 1]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # ring ABA has no junctions
@@ -200,13 +201,13 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[0, 0], [1, 0], [1, 1], [0, 0]],
             }
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # ring AA is not a proper polygon geometry.
     def test_single_ring_AA(self):
         data = {"aa": {"type": "Polygon", "coordinates": [[0, 0], [0, 0]]}}
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when a new line DEC shares its end with an old line ABC, there is no junction at C
@@ -215,7 +216,7 @@ class TestJoin(unittest.TestCase):
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "dec": {"type": "LineString", "coordinates": [[0, 1], [1, 1], [2, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when a new line ABC extends an old line AB, there is a junction at B
@@ -224,7 +225,7 @@ class TestJoin(unittest.TestCase):
             "ab": {"type": "LineString", "coordinates": [[0, 0], [1, 0]]},
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(
             geometry.MultiPoint(topo["junctions"]).equals(
             geometry.MultiPoint([(0.0, 0.0), (1.0, 0.0)])),
@@ -236,7 +237,7 @@ class TestJoin(unittest.TestCase):
             "ba": {"type": "LineString", "coordinates": [[1, 0], [0, 0]]},
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(
             geometry.MultiPoint(topo["junctions"]).equals(
             geometry.MultiPoint([(1.0, 0.0), (0.0, 0.0)])))
@@ -248,7 +249,7 @@ class TestJoin(unittest.TestCase):
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "bc": {"type": "LineString", "coordinates": [[1, 0], [2, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(
             geometry.MultiPoint(topo["junctions"]).equals(
                 geometry.MultiPoint([(1.0, 0.0), (0.0, 0.0), (2.0, 0.0)])
@@ -262,7 +263,7 @@ class TestJoin(unittest.TestCase):
             "cba": {"type": "LineString", "coordinates": [[2, 0], [1, 0], [0, 0]]},
             "bc": {"type": "LineString", "coordinates": [[1, 0], [2, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(
             geometry.MultiPoint(topo["junctions"]).equals(
                 geometry.MultiPoint([(2.0, 0.0), (1.0, 0.0)])
@@ -275,7 +276,7 @@ class TestJoin(unittest.TestCase):
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "abd": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [3, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(
             geometry.MultiPoint(topo["junctions"]).equals(
                 geometry.MultiPoint([(0.0, 0.0), (2.0, 0.0)])
@@ -289,7 +290,7 @@ class TestJoin(unittest.TestCase):
             "cba": {"type": "LineString", "coordinates": [[2, 0], [1, 0], [0, 0]]},
             "abd": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [3, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(
             geometry.MultiPoint(topo["junctions"]).equals(
                 geometry.MultiPoint([(2.0, 0.0), (0.0, 0.0)])
@@ -302,7 +303,7 @@ class TestJoin(unittest.TestCase):
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "dbc": {"type": "LineString", "coordinates": [[3, 0], [1, 0], [2, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(
             geometry.MultiPoint(topo["junctions"]).equals(
                 geometry.MultiPoint([(1.0, 0.0), (2.0, 0.0), (3.0, 0.0), (0.0, 0.0)])
@@ -315,7 +316,7 @@ class TestJoin(unittest.TestCase):
             "cba": {"type": "LineString", "coordinates": [[2, 0], [1, 0], [0, 0]]},
             "dbc": {"type": "LineString", "coordinates": [[3, 0], [1, 0], [2, 0]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertTrue(
             geometry.MultiPoint(topo["junctions"]).equals(
                 geometry.MultiPoint([(2.0, 0.0), (1.0, 0.0), (3.0, 0.0)])
@@ -329,7 +330,7 @@ class TestJoin(unittest.TestCase):
             "abc": {"type": "LineString", "coordinates": [[0, 0], [1, 0], [2, 0]]},
             "dbe": {"type": "LineString", "coordinates": [[0, 1], [1, 0], [2, 1]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when a new line ABDE skips a point with an old line ABCDE, there is a no junction
@@ -344,7 +345,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[0, 0], [1, 0], [3, 0], [4, 0]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when a new line ABDE skips a point with a reversed old line EDCBA, there is 
@@ -360,7 +361,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[0, 0], [1, 0], [3, 0], [4, 0]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when a line ABCDBE self-intersects with its middle, there are no junctions
@@ -371,7 +372,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[0, 0], [1, 0], [2, 0], [3, 0], [1, 0], [4, 0]],
             }
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when a line ABACD self-intersects with its start, there are no junctions
@@ -382,7 +383,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[0, 0], [1, 0], [0, 0], [3, 0], [4, 0]],
             }
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when a line ABCDBD self-intersects with its end, there are no junctions
@@ -393,7 +394,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[0, 0], [1, 0], [4, 0], [3, 0], [4, 0]],
             }
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when an old line ABCDBE self-intersects and shares a point B, there is 
@@ -406,7 +407,7 @@ class TestJoin(unittest.TestCase):
             },
             "fbg": {"type": "LineString", "coordinates": [[0, 1], [1, 0], [2, 1]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when a line ABCA is closed, there is no junction at A
@@ -417,7 +418,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[0, 0], [1, 0], [0, 1], [0, 0]],
             }
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # when a ring ABCA is closed, there are no junctions
@@ -428,7 +429,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[0, 0], [1, 0], [0, 1], [0, 0]]],
             }
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # exact duplicate rings ABCA & ABCA share the arc ABCA, but contain no junctions
@@ -443,7 +444,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[0, 0], [1, 0], [0, 1], [0, 0]]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # reversed duplicate rings ABCA & ACBA share the arc ABCA, but contain no juctions
@@ -458,7 +459,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[0, 0], [0, 1], [1, 0], [0, 0]]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # coincident rings ABCA & BCAB share the arc BCAB, but contain no junctinos
@@ -475,7 +476,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[1, 0], [0, 1], [0, 0], [1, 0]]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # coincident rings ABCA & BACB share the arc BCAB, but contain no junctions
@@ -490,7 +491,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[1, 0], [0, 0], [0, 1], [1, 0]]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # coincident rings ABCA & DBED share the point B, but is no junction
@@ -505,7 +506,7 @@ class TestJoin(unittest.TestCase):
                 "coordinates": [[[2, 1], [1, 0], [2, 2], [2, 1]]],
             },
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # coincident ring ABCA & line DBE share the point B
@@ -517,17 +518,17 @@ class TestJoin(unittest.TestCase):
             },
             "dbe": {"type": "LineString", "coordinates": [[2, 1], [1, 0], [2, 2]]},
         }
-        topo = topojson.Join(data).to_dict()
+        topo = Join(data).to_dict()
         self.assertListEqual(topo["junctions"], [])
 
     # should keep junctions from partly shared paths
     # this test was added since the a shared path of ABC and another shared path of ABD
     # only kept the junctions at A, C and D, but not at B.
     def test_shared_junctions_in_shared_paths(self):
-        data = geopandas.read_file(geopandas.datasets.get_path('naturalearth_lowres'))
-        data = data[(data.name == 'Togo') | (data.name == 'Benin') | 
-            (data.name == 'Burkina Faso')]
-        topo = topojson.Join(data).to_dict()
+        data = geopandas.read_file(geopandas.datasets.get_path("naturalearth_lowres"))
+        data = data[(data.name == "Togo") | (data.name == "Benin") | 
+            (data.name == "Burkina Faso")]
+        topo = Join(data).to_dict()
         self.assertEqual(len(topo["junctions"]), 6)
 
     # this test was added since a shared path can be detected of two linestrings where 
@@ -543,19 +544,19 @@ class TestJoin(unittest.TestCase):
             | (data.name == "Ethiopia")
             | (data.name == "Sudan")
         ]
-        topo = topojson.Join(data).to_dict()
-        self.assertEqual(len(topo['junctions']), 6)  
+        topo = Join(data).to_dict()
+        self.assertEqual(len(topo["junctions"]), 6)  
 
     def test_non_noded_interesection(self):
         data = geopandas.read_file(geopandas.datasets.get_path("naturalearth_lowres"))
-        topo = topojson.Join(data).to_dict()#, quant_factor=1e6)
-        self.assertEqual(len(topo['junctions']), 321)   
+        topo = Join(data).to_dict()#, quant_factor=1e6)
+        self.assertEqual(len(topo["junctions"]), 321)   
 
     def test_super_function_extract(self):
         data = geometry.GeometryCollection([
             geometry.Polygon([[0, 0], [1, 0], [1, 1], [0, 1], [0, 0]]), 
             geometry.Polygon([[1, 0], [2, 0], [2, 1], [1, 1], [1, 0]])
         ])   
-        topo = topojson.Join(data).to_dict()  
+        topo = Join(data).to_dict()  
         self.assertEqual(list(topo.keys()), 
-        ['type', 'linestrings', 'bookkeeping_geoms', 'objects', 'junctions'])
+        ["type", "linestrings", "bookkeeping_geoms", "objects", "options", "junctions"])
